+++ nanopolish_profile_hmm.h
@@ -20,7 +20,7 @@
 //
 
 // Calculate the probability of the nanopore events given the consensus sequence
-double profile_hmm_score(const std::string& consensus, const HMMInputData& data);
+float profile_hmm_score(const std::string& consensus, const HMMInputData& data);
 
 // Run viterbi to align events to kmers
 std::vector<AlignmentState> profile_hmm_align(const std::string& sequence, const HMMInputData& data);
@@ -30,13 +30,13 @@ std::vector<AlignmentState> profile_hmm_align(const std::string& sequence, const
 //
 
 // Initialize the forward algorithm
-void profile_hmm_forward_initialize(DoubleMatrix& fm);
+void profile_hmm_forward_initialize(FloatMatrix& fm);
 
 // Terminate the forward algorithm
-double profile_hmm_forward_terminate(const DoubleMatrix& fm, uint32_t row);
+float profile_hmm_forward_terminate(const FloatMatrix& fm, uint32_t row);
 
 // Fill in the forward matrix
-double profile_hmm_forward_fill(DoubleMatrix& fm, // forward matrix
+float profile_hmm_forward_fill(FloatMatrix& fm, // forward matrix
                                 const char* sequence,
                                 const HMMInputData& data,
                                 uint32_t e_start);
@@ -47,10 +47,10 @@ double profile_hmm_forward_fill(DoubleMatrix& fm, // forward matrix
 // 
 
 // initialize viterbi
-void profile_hmm_viterbi_initialize(DoubleMatrix& m);
+void profile_hmm_viterbi_initialize(FloatMatrix& m);
 
 // fill in the score matrix and the backtrack matrix
-void profile_hmm_viterbi_fill(DoubleMatrix& vm, // viterbi matrix
+void profile_hmm_viterbi_fill(FloatMatrix& vm, // viterbi matrix
                               UInt8Matrix& bm, // backtrack matrix
                               const char* sequence,
                               const HMMInputData& data,
